requirements.txt
Django>=4.0,<5.0
gunicorn>=20.0,<21.0
psycopg2-binary<3.0,>=2.9
djangorestframework==3.14.0
PyJWT==2.6.0
django-cors-headers
